+++ packages/extra/src/Querying/ConditionParsers/Drupal/DrupalConditionFactoryInterface.php
@@ -9,21 +9,23 @@ use EDT\Querying\Contracts\PathsBasedInterface;
 /**
  * @template TCondition of \EDT\Querying\Contracts\PathsBasedInterface
  */
-interface OperatorProviderInterface
+interface DrupalConditionFactoryInterface
 {
     /**
      * Returns all Drupal operator names supported by this instance.
      *
      * @return list<non-empty-string>
      */
-    public function getAllOperatorNames(): array;
+    public function getSupportedOperators(): array;
 
     /**
-     * @param non-empty-string                       $operatorName
-     * @param mixed|null                             $value
+     * @param non-empty-string                 $operatorName
+     * @param mixed|null                       $value
      * @param non-empty-list<non-empty-string> $path
      *
      * @return TCondition
+     *
+     * @throws DrupalFilterException if the given operator name is not supported
      */
-    public function createOperator(string $operatorName, $value, array $path): PathsBasedInterface;
+    public function createCondition(string $operatorName, $value, array $path): PathsBasedInterface;
 }